trailing fmt, moved backtestoutput into perf
