+++ src/types/mod.rs
@@ -1,7 +1,7 @@
 use itertools::Itertools;
-use std::{collections::HashMap, ops::Add};
 use std::hash::Hash;
 use std::ops::Deref;
+use std::{collections::HashMap, ops::Add};
 use time::{format_description, Date, OffsetDateTime};
 
 ///Defines a set of base types that are used by multiple components.
@@ -381,17 +381,4 @@ pub struct StrategySnapshot {
     pub date: DateTime,
     pub portfolio_value: CashValue,
     pub net_cash_flow: CashValue,
-}
-
-///Performance output from a single backtest run.
-#[derive(Clone, Debug)]
-pub struct BacktestOutput {
-    pub ret: f64,
-    pub cagr: f64,
-    pub vol: f64,
-    pub mdd: f64,
-    pub sharpe: f64,
-    pub values: Vec<f64>,
-    pub returns: Vec<f64>,
-    pub dates: Vec<i64>,
 }
